requirements.txt
python-binance
pandas
python-telegram-bot